requirements.txt
<<<<<<< HEAD
astroid==2.11.2
certifi==2020.11.8
cftime==1.3.0
click==7.1.2
colorama==0.4.4
configparser==5.0.1
cycler==0.10.0
dill==0.3.4
GDAL==3.1.2
isort==5.10.1
Jinja2==2.11.2
kiwisolver==1.3.1
lazy-object-proxy==1.7.1
MarkupSafe==1.1.1
matplotlib==3.3.3
mccabe==0.7.0
mpld3==0.5.1
netCDF4==1.5.4
numpy==1.21.0
pandas==1.1.4
pathlib==1.0.1
Pillow==9.0.1
platformdirs==2.5.2
PyCRS==1.0.2
pyparsing==2.4.7
pyproj==3.0.0.post1
python-dateutil==2.8.1
pytz==2020.4
scipy==1.5.4
Shapely==1.7.1
six==1.15.0
tomli==2.0.1
typed-ast==1.5.3
typing-extensions==4.2.0
waterpy==0.2
wrapt==1.14.0
=======
astroid==2.11.2
certifi==2022.12.7
cftime==1.3.0
click==7.1.2
colorama==0.4.4
configparser==5.0.1
cycler==0.10.0
dill==0.3.4
GDAL==3.1.2
isort==5.10.1
Jinja2==2.11.3
kiwisolver==1.3.1
lazy-object-proxy==1.7.1
MarkupSafe==1.1.1
matplotlib==3.3.3
mccabe==0.7.0
mpld3==0.5.1
netCDF4==1.5.4
numpy==1.22.0
pandas==1.1.4
pathlib==1.0.1
Pillow==9.3.0
platformdirs==2.5.2
PyCRS==1.0.2
pyparsing==2.4.7
pyproj==3.0.0.post1
python-dateutil==2.8.1
pytz==2020.4
scipy==1.5.4
Shapely==1.7.1
six==1.15.0
tomli==2.0.1
typed-ast==1.5.3
typing-extensions==4.2.0
waterpy==0.2
wrapt==1.14.0
>>>>>>> b21aa33c4be91bdba14a980db6e5ac9e0ce56529
